requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-openai
langchain-google-genai
langchain-groq
langchain-mistralai
python-dotenv
faiss-cpu
tiktoken
fastapi
uvicorn
requests
langchain_huggingface

# Tavily Search Tool
tavily-python>=0.3.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# Environment variables
python-dotenv>=1.0.0

# Async support (important for agents/tools)
aiohttp>=3.9.0

# Data handling (optional but useful in pipelines)
pandas>=2.0.0

# Token counting / text utils
tiktoken>=0.6.0

# Logging / debugging (optional but recommended)
rich>=13.7.0

# Retry + robustness
tenacity>=8.2.0

# JSON handling improvements
orjson>=3.9.0

# Typing support (helps in bigger projects)
pydantic>=2.5.0

# Optional (for advanced parsing / HTML cleanup)
html5lib>=1.1


# LLM Providers
openai
mistralai

# Vector Database
chromadb

# Embedding Models
sentence-transformers

# Document Loaders dependencies
pypdf
unstructured
beautifulsoup4
lxml

# Text Processing
tiktoken

# Environment Variables
python-dotenv

# Data Handling
pandas
numpy

# API / App Layer (optional but common in RAG apps)
fastapi
uvicorn

# UI (optional)
streamlit

# Utilities
tqdm
requests